Started moving around code.
Splitting up the mapper code base into more logical pieces.
